Renaming to generic form for simpler test unmocking
