+++ graphql/api/applications/resolvers.js
