+++ graphql/api/groups/schema.js
@@ -1,6 +1,6 @@
-import Applications from '../applications/ApplicationSchema';
-import Frameworks from '../frameworks/FrameworkSchema';
-import Tasks from '../tasks/TaskSchema';
+import Applications from '../applications/schema';
+import Frameworks from '../frameworks/schema';
+import Tasks from '../tasks/schema';
 
 const Groups = `
   # Union which resolves to one of Group, Application or Framework
